Debate2Decision/backend/requirements.txt
fastapi
uvicorn
websockets
httpx
pydantic
python-dotenv
edge-tts
gradio_client
Pillow
